requirements.txt
PyQt6>=6.4.0
zhipuai>=2.0.0
edge-tts>=6.1.0
playsound
pygame>=2.1.0
requests>=2.28.0
openai>=1.0.0
Pillow>=9.0.0
stegano
pyyaml